splitted InitOptions and removed types folder for structures
