+++ src/commands/base/GuildCommand.ts
@@ -1,8 +1,25 @@
 import { CommandManager } from "../../structures/CommandManager";
-import { FunctionCommand } from "./FunctionCommand";
-import { GuildCommandInit } from "../types/InitOptions";
+import { FunctionCommand, FunctionCommandInit } from "./FunctionCommand";
 import { CommandType } from "../types/commands";
 import { InputManager } from "../../structures/InputManager";
+
+/**
+ * Initialization options of base guild-scoped command
+ * @interface
+ * @extends {FunctionCommandInit}
+ */
+export interface GuildCommandInit extends FunctionCommandInit {
+    /**
+     * Whether this command should be callable using private messages with bot
+     * @type {?boolean}
+     */
+    dm?: boolean;
+    /**
+     * List of Guild IDs in which the command can be called
+     * @type {?Array<string>}
+     */
+    guilds?: string[];
+}
 
 /**
  * Guild-scoped executable command
