+++ src/structures/apiTypes.ts
